Refactor http module
Blocking and streaming http responses now live in a separate classes.
The blocking response is implemented using the streaming response. Each
streaming response has its own curl and curl_multi handle.

The streaming response has to be explicitly closed. The preferred way is
to use it as a context manager which ensures that it is closed when
exception is raised in the middle of reading the stream. Because it
doesn't fit into our current API, the class also tries to free the curl
resources when it finishes reading the http stream and also when it's
gargabe collected.